% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geuv.R
\name{getGeuvRNA}
\alias{getGeuvRNA}
\title{obtain a summarized experiment with GEUVADIS (Coriell) RNA-seq quantifications}
\usage{
getGeuvRNA()
}
\description{
obtain a summarized experiment with GEUVADIS (Coriell) RNA-seq quantifications
}
\examples{
requireNamespace(SummarizedExperiment)
gg = getGeuvRNA()
gg
SummarizedExperiment::assay(gg[1:4,1:5])
}
