% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igsr_pops}
\alias{igsr_pops}
\title{table with metadata on populations for IGSR
docType data}
\format{
An object of class \code{data.frame} with 212 rows and 11 columns.
}
\usage{
igsr_pops
}
\description{
table with metadata on populations for IGSR
docType data
}
\examples{
data(igsr_pops)
head(igsr_pops)
}
\keyword{datasets}
