% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geuv.R
\name{install_early_t2t_txdb}
\alias{install_early_t2t_txdb}
\title{install early version of T2T TxDb, do nothing if already available}
\usage{
install_early_t2t_txdb()
}
\description{
install early version of T2T TxDb, do nothing if already available
}
