% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_190kloci}
\alias{pc_190kloci}
\title{10 PCs for 3202 samples based on .05 random sample from chr17}
\format{
matrix 3202 x 10
}
\usage{
pc_190kloci
}
\description{
10 PCs for 3202 samples based on .05 random sample from chr17
}
\examples{
data(pc_190kloci)
head(pc_190kloci)
}
\keyword{datasets}
