% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_3kloci}
\alias{pc_3kloci}
\title{10 PCs for 3202 samples based on .001 random sample from chr17}
\format{
matrix 3202 x 10
}
\usage{
pc_3kloci
}
\description{
10 PCs for 3202 samples based on .001 random sample from chr17
}
\examples{
data(pc_3kloci)
head(pc_3kloci)
}
\keyword{datasets}
