% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_990kloci}
\alias{pc_990kloci}
\title{10 PCs for 3202 samples based on .25 random sample from chr17}
\format{
matrix 3202 x 10
}
\usage{
pc_990kloci
}
\description{
10 PCs for 3202 samples based on .25 random sample from chr17
}
\examples{
data(pc_990kloci)
head(pc_990kloci)
}
\keyword{datasets}
