% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewStrat.R
\name{strat2d}
\alias{strat2d}
\title{static visualization of PCA pairs}
\usage{
strat2d(
  pcmat,
  pcinds = 5:8,
  superpop = "AFR",
  GGally = FALSE,
  superOnly = FALSE
)
}
\arguments{
\item{pcmat}{matrix with columns PC1...PCn}

\item{pcinds}{numeric vector with suffixes from colnames(pcmat) to select for visualization}

\item{superpop}{character(1) three letter superpopulation code, see `data(igsr_pops)`}

\item{GGally}{logical(1) if TRUE, use `GGally::ggpairs` to visualize, otherwise `pairs()`}

\item{superOnly}{logical(1) defaults to FALSE; if TRUE, do not limit to one superpopulation}
}
\value{
Side effects of plotting with either `pairs` or `GGally`
}
\description{
static visualization of PCA pairs
}
\examples{
data(pc_190kloci)
strat2d(pc_190kloci)
}
