% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewStrat.R
\name{stratapp}
\alias{stratapp}
\title{multipanel app to view stratification patterns in various populations}
\usage{
stratapp()
}
\description{
multipanel app to view stratification patterns in various populations
}
