% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewStrat.R
\name{viewStrat1KG}
\alias{viewStrat1KG}
\title{visualize PCA colored by populations for a selected SuperPopulation}
\usage{
viewStrat1KG(
  pcs,
  selectedPCs = 1:4,
  superpop = "AFR",
  superOnly = FALSE,
  dropNAs = TRUE
)
}
\arguments{
\item{pcs}{matrix of principal components, assumed to have colnames PC[1,...] typically PC1-PC10}

\item{selectedPCs}{numeric(), values typically drawn from 1:10}

\item{superpop}{character(1)}

\item{superOnly}{logical(1) defaults to FALSE; if TRUE, do not limit to one superpopulation}

\item{dropNAs}{logical(1) defaults to TRUE}
}
\description{
visualize PCA colored by populations for a selected SuperPopulation
}
\note{
Population and SuperPopulation definitions are at `https://www.internationalgenome.org/data-portal/population`,
mapping from sample to population/superpopulation is available via `BiocHail::path_1kg_annotations`.
}
